# docs/plan/syncnode/openapi.yaml
openapi: 3.0.3
info:
  title: CubeFS Syncnode — Complete Admin API
  version: "2.0.0"
  description: |
    Complete HTTP admin surface for the CubeFS sync subsystem.

    **Two servers, two surfaces:**

    | Server | Default Port | Surface |
    |--------|-------------|---------|
    | Master | 17010 | All sync management APIs: node registration, rule CRUD, task observability, direct dispatch |
    | Syncnode | 17911 | Local diagnostics only: version, stat, reload |

    **Authentication:**
    All master sync endpoints require `Authorization: Bearer <syncAdminToken>` (or
    `X-Sync-Token: <token>`). The token is configured as `syncAdminToken` in
    `master.json`. An empty value disables the check (dev/test mode).

    Syncnode local endpoints share the same token scheme but are only
    reachable on the node itself.

    **Response envelope (master + syncnode):**
    ```json
    { "code": 0, "msg": "OK", "data": <payload> }
    ```
    Non-zero `code` signals an error; `data` is null. HTTP status is
    always 200 — check `code` in the body.

    **Common error codes:**
    | code | meaning |
    |------|---------|
    | 0 | success |
    | 7 | ErrCodeParamError — missing or malformed param / body |
    | 10 | ErrCodePersistenceByRaft — raft submit failed; safe to retry |
    | 11 | ErrCodeInternalError — not found, conflict, or other server error |

  contact:
    name: CubeFS Sync Team
servers:
  - url: "http://{host}:{masterPort}"
...
    variables:
      host:
...
security:
  - BearerAuth: []
  - SyncTokenAuth: []

# ────────────────────────────────────────────────────────────────────────────
# Master: SyncNode registration + listing
# ────────────────────────────────────────────────────────────────────────────
paths:
  /syncNode/add:
    post:
      tags: [node-registration]
      summary: 注册 syncnode（syncnode 启动时自调用）
      description: |
        Syncnode self-registration. Mirrors `/lcNode/add`. Idempotent on addr —
        calling again for an already-registered addr returns the existing nodeID.
        Master uses the addr for task dispatch and heartbeat tracking.
      operationId: addSyncNode
      parameters:
        - name: addr
          in: query
          required: true
          description: "`host:port` of the registering syncnode"
          schema:
            type: string
            example: "10.89.6.10:17911"
      responses:
        "200":
          description: OK — returns allocated nodeID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: integer
                        format: int64
                        description: Master-allocated numeric node ID
                        example: 42
...
  /syncRule/create:
    post:
      tags: [rule-mgmt]
      summary: 创建同步规则
      description: |
        Body is a **flat `SyncRuleConfig` JSON object** — not the outer
        `SyncRule` wrapper. Server controls `state`, `createdAt`,
        `updatedAt`, and `lastRun*` fields.

        Validation rules (400 on violation):
        - `id` required; must not contain `/`.
        - `type` required.
        - `shardingStrategy` must be one of: `""`, `"hash"`, `"prefix"`,
          `"auto"`.
        - When `shardingStrategy == "prefix"`, `shardPrefixes` must be
          non-empty.

        Newly created rules are always `active`. To start paused, create
        then call `/syncRule/pause`.
      operationId: createSyncRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRuleConfig"
            example:
              id: "backup-cfs-to-s3"
              type: "sync"
              schedule: "0 2 * * *"
              src:
                kind: "cfs"
                vol: "myvol"
                path: "/data"
              dst:
                kind: "s3"
                endpoint: "https://s3.ap-east-1.amazonaws.com"
                region: "ap-east-1"
                bucket: "my-backup"
                prefix: "daily/"
              parallelism: 4
              bandwidthLimitMBps: 200
      responses:
        "200":
          description: Created — returns full SyncRule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SyncRule"
  /syncRule/update:
    post:
      tags: [rule-mgmt]
      summary: 更新同步规则配置
      description: |
        Body is the same flat `SyncRuleConfig` shape as create. The rule
        must already exist (identified by `id`). Server merges caller's
        `config` over the existing record and refreshes `updatedAt`.
        Server-owned fields (`state`, `createdAt`, `lastRun*`) are
        preserved.
      operationId: updateSyncRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRuleConfig"
      responses:
        "200":
          description: Updated — returns full SyncRule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SyncRule"
  /syncRule/delete:
    post:
      tags: [rule-mgmt]
      summary: 删除同步规则
      description: |
        Raft-deletes the rule record and removes it from the cron
        scheduler. In-flight tasks are NOT cancelled automatically —
        cancel them first via `/syncTask/cancel` if needed.
      operationId: deleteSyncRule
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          example: "backup-cfs-to-s3"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
  /syncRule/pause:
    post:
      tags: [rule-mgmt]
      summary: 暂停规则（停止调度触发）
      description: |
        Transitions the rule to `paused`. Cron scheduler stops firing this
        rule until resumed. In-flight tasks are unaffected — they continue
        to run.

        Idempotent: calling on an already-paused rule returns success.
      operationId: pauseSyncRule
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused — returns updated SyncRule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SyncRule"
  /syncRule/resume:
    post:
      tags: [rule-mgmt]
      summary: 恢复规则调度
      description: |
        Transitions the rule from `paused` or `degraded` back to `active`.
        Cron scheduler resumes firing the rule on its next tick.

        Idempotent: calling on an already-active rule returns success.
      operationId: resumeSyncRule
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed — returns updated SyncRule
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SyncRule"
  /syncRule/list:
    get:
      tags: [rule-mgmt]
      summary: 列出同步规则
      description: |
        Returns all cached rules. Optional `state` filter. Output is
        unordered — sort client-side as needed.
      operationId: listSyncRules
      parameters:
        - name: state
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/SyncRuleState"
          description: When omitted, all rules regardless of state are returned
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/SyncRule"
  /syncRule/get:
    get:
      tags: [rule-mgmt]
      summary: 获取单条规则
      operationId: getSyncRule
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        $ref: "#/components/schemas/SyncRule"
  /syncRule/trigger:
    post:
      tags: [rule-mgmt]
      summary: 立即触发规则执行（跳过 cron 等待）
      description: |
        Fires the rule's dispatch path immediately — same code path as the
        cron callback. Generates a fresh taskID (`<ruleID>/<nanoseconds>`).
        Useful for ops scripts and integration tests that don't want to
        wait for the next scheduled tick.

        Rule must be in `active` state — paused or degraded rules are
        rejected (400).

        Concurrent calls produce distinct taskIDs; the dispatcher's ledger
        de-duplicates against in-flight tasks so concurrent triggers don't
        collide.
      operationId: triggerSyncRule
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          example: "backup-cfs-to-s3"
      responses:
        "200":
          description: Dispatched — returns ruleID + new taskID
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Response"
                  - properties:
                      data:
                        type: object
                        properties:
                          ruleID:
                            type: string
                          taskID:
                            type: string
                            example: "backup-cfs-to-s3/1715798400000000000"

# ────────────────────────────────────────────────────────────────────────────
# Master: SyncTask observability + lifecycle
# ────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer <syncAdminToken>`"
    SyncTokenAuth:
      type: apiKey
      in: header
      name: X-Sync-Token
      description: Alternative to Bearer; same token value from `syncAdminToken` in master.json
  schemas:
...
    Response:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 0
          description: "0 = success. Non-zero = error (check msg). HTTP status is always 200."
        msg:
          type: string
          example: "OK"
        data:
          nullable: true
          description: Payload on success; null on error.

    # ── SyncRule ───────────────────────────────────────────────────────────
    SyncRuleState:
      type: string
      enum: [active, paused, degraded]
      description: |
        - `active`: rule fires on schedule
        - `paused`: schedule suspended by operator
        - `degraded`: auto-paused by master after repeated task failures
    SyncEndpointConfig:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [cfs, s3, local]
          description: Backend type. Determines which other fields are used.
        # cfs fields
        vol:
          type: string
          description: CubeFS volume name (cfs only)
        path:
          type: string
          description: Path within the volume / local filesystem
        # s3 fields
        bucket:
          type: string
        prefix:
          type: string
        endpoint:
          type: string
          example: "https://s3.ap-east-1.amazonaws.com"
        region:
          type: string
        storageClass:
          type: string
        accessKeyEnv:
          type: string
          description: Name of env var holding the S3 access key (overrides global s3Defaults)
        secretKeyEnv:
          type: string
          description: Name of env var holding the S3 secret key
        insecureSkipTLS:
          type: boolean
          default: false
          description: Disable TLS certificate verification (dev/test only)
        usePathStyle:
          type: boolean
          default: false
          description: Force S3 path-style addressing (required for Ceph RGW, ByteDance TOS, etc.)
        # local fields
        bufferSizeKiB:
          type: integer
        concurrency:
          type: integer
        directIO:
          type: boolean
        fadviseSequential:
          type: boolean
    SyncFilterConfig:
      type: object
      properties:
        include:
          type: array
          items:
            type: string
          description: Glob patterns — only matching files are synced
        exclude:
          type: array
          items:
            type: string
          description: Glob patterns — matching files are skipped
        minSize:
          type: string
          description: "Human-readable size threshold, e.g. \"1MB\""
        maxSize:
          type: string
          description: "Human-readable size threshold, e.g. \"10GB\""
        minAge:
          type: string
          description: "Duration string, e.g. \"30s\", \"1h\""
        maxAge:
          type: string
    SyncRetentionConfig:
      type: object
      properties:
        pattern:
          type: string
          description: Glob pattern selecting destination files subject to retention
        keepLast:
          type: integer
          description: Keep the N most recent matching objects
        keepWithin:
          type: string
          description: "Keep objects created within this duration, e.g. \"7d\""
    SyncRuleConfig:
      type: object
      required: [id, type]
      description: |
        Wire body for `/syncRule/create` and `/syncRule/update`.
        Server-owned fields (`state`, `createdAt`, `updatedAt`, `lastRun*`)
        are NOT accepted in the body and are always overwritten server-side.
      properties:
        id:
          type: string
          description: Stable unique identifier. Must not contain `/`.
          example: "backup-cfs-to-s3"
        type:
          type: string
          description: Task type understood by the syncnode executor (e.g. `"sync"`)
          example: "sync"
        schedule:
          type: string
          description: Cron expression (5-field, UTC). Leave empty for manual-only rules.
          example: "0 2 * * *"
        src:
          $ref: "#/components/schemas/SyncEndpointConfig"
        dst:
          $ref: "#/components/schemas/SyncEndpointConfig"
        filter:
          $ref: "#/components/schemas/SyncFilterConfig"
        retention:
          $ref: "#/components/schemas/SyncRetentionConfig"
        afterCopy:
          type: string
          enum: ["", "delete-src", "archive-src"]
          description: Action on source file after successful copy
        downloadStrategy:
          type: string
          description: Strategy hint for the executor (backend-specific)
        onMismatch:
          type: string
          enum: ["", "overwrite", "skip", "fail"]
          description: Behaviour when src and dst checksums differ
        sampleStrategy:
          type: string
          description: Verification sampling strategy
        sampleRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Fraction of files to verify (0.0–1.0)
        bandwidthLimitMBps:
          type: integer
          description: Per-node egress bandwidth ceiling in MB/s (0 = unlimited)
        aggregateBandwidthLimitMBps:
          type: integer
          description: Cluster-wide egress ceiling for this rule in MB/s (0 = unlimited)
        parallelism:
          type: integer
          description: Number of concurrent file transfer workers per task execution
        shardingStrategy:
          type: string
          enum: ["", "hash", "prefix", "auto"]
          description: |
            Fan-out strategy when master dispatches this rule across multiple syncnodes:
            - `""` / `"hash"`: FNV-1a hash on object key (default; even distribution)
            - `"prefix"`: use `shardPrefixes` literally; len(shardPrefixes) defines shard count
            - `"auto"`: master probes backend top-level prefixes at fire time
        shardPrefixes:
          type: array
          items:
            type: string
          description: Explicit partition prefixes for `shardingStrategy=prefix`
    SyncRule:
      type: object
      description: Full runtime view of a sync rule as stored and returned by master.
      properties:
        config:
          $ref: "#/components/schemas/SyncRuleConfig"
        state:
          $ref: "#/components/schemas/SyncRuleState"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastRunAt:
          type: string
          format: date-time
          description: Zero / omitted until the first task completes
        lastRunStatus:
          type: string
          enum: ["", "done", "failed", "cancelled"]
        lastRunError:
          type: string
          description: Non-empty only when `lastRunStatus == "failed"`

    # ── SyncNode ───────────────────────────────────────────────────────────
...